Add pyBullet driver for single finger
Add a C++ wrapper for the Finger/TriFinger simulation implementing `robot_interfaces::RobotDriver` as well as a function and Python bindings to create a "simulation backend" for the robot_interfaces pipeline.

Add a demo `demo_robot_interfaces.py` for testing it.

Add some helper functions to `SimFinger` so the driver can easily access the functionality it needs.