Store action log and add replay script to verify
Add method to `TriFingerPlatform` to log all actions and store it to a
JSON file together with initial and final object state.

Add a script `replay_action_log.py` which replays these actions in a new
simulation and verifies that the resulting final object state matches
with the one of the log.