Adjust bounds for cube position
The arena boundary has a diameter of around 56cm at the top, so assuming
a radius of 30cm from the center should cover all cube positions inside
the boundary.

The height of 30cm seems to be good already, as this roughly corresponds
to the height of the fingers and it should not be possible to lift the
object higher than that.