Simulate raw camera data using Bayer pattern
The cameras on the real robot provide their images in raw format (Bayer
BG) and need to be converted by the user.  To get the same behaviour
from the simulation, convert the rendered RGB image to a Bayer pattern
by keeping only one channel from each pixel, according to the pattern.

Also reduce the default image resolution to 270x270 as on the real
robot.