Remove machine-dependent path to xacro from urdf files
To avoid unnecessary diffs in the future, shorten the path to the
original xacro file in the comment header of the urdf files to only the
part inside the robot_properties_fingers package.
The part above that is different for each user/workspace so it causes
diffs even on files which were not changed.  On the other hand, this
part hardly adds any value.