Port update_robot_properties.sh to ROS2
There is no `rospack find` anymore.
As there doesn't seem to be a replacement to get the path to the package
source directory, use the location of the update script itself to find
the trifinger_simulation source.
For robot_properties_fingers use `ros2 pkg prefix` which provides the
path to the install space.