Add CalibratedCamera and refactor
Add a `CalibratedCamera` class which is a replacement for `Camera` that
renders more realistic images.  It takes the full camera calibration
parameters (camera matrix, distortion coefficients and pose) into
consideration.  So when fed with parameters of an actual robot, the
rendered images are much closer to the real ones (w.r.t. pose, field of
view, and distortion) than the ones from the old `Camera` class.

For this, it is necessary to manually construct the "projection matrix"
needed by pyBullet as pyBullet's `computeProjectionMatrix` makes some
assumptions (fx = fy, x0, y0 are exactly in the centre) which are not
actually true for our cameras.
The distortion is achieved by first rendering an undistorted image and
then distorting it.  To avoid empty areas at the edges the image is
first rendered at a slightly larger size as desired and then cropped
after distorting.

Refactoring:

To make the code a bit more general, add generic `CameraArray` class
which is simply a wrapper around a list of arbitrarily many Cameras.
Use this as base class for the old `TriFingerCameras` class and also
provide a factory function creating a `CameraArray` based on camera
calibration files.

Handle multiple pybullet clients the same way as everywhere else in
`Camera` (i.e. add an optional argument `pybullet_client_id` which is
passed on).