Export cube_v2.obj with proper orientation
One need's to take care to set the orientation settings correctly when
exporting the obj file in Blender as unfortunately it does not simply
preserve the orientation of Blender...

The correct settings seem to be "Z up, Y forward".  It can easily be
checked by opening the obj file in MeshLab and visualising the axis.