Subpackage move_cube_on_trajectory with evaluation in __main__.py
Convert the move_cube_on_trajectory module to a subpackage (move the
code from the old module to `__init__.py` of the subpackage, so it can
be used in the same way).

Add all the scripts for evaluation and replay as modules and provide a
`__main__.py` to call them in a convenient way.  These modules are all
adapted versions of the corresponding scripts that already exist for the
move_cube task.