Also use cached camera observation for t + 1
Always using `_get_current_camera_observation()` results in a very
different behaviour as the lower camera update rate is not considered in
this case.  Instead simply use the cached observation as for step `t`.
This will result in a slight offset of the camera update but given the
already pretty slow camera rate (compared to the robot rate), this
should hopefully not cause much trouble.