Move move_cube evaluation scripts to move_cube package
Move the scripts from scripts/ to the package and use `__main__.py` to
execute them.  This follows the same pattern as with the new
`move_cube_on_trajectory` task, thus allowing us to have equivalent
structures for different tasks without causing confusion (with the files
in the `scripts/` directory, it is not directly clear which task they
belong to).