Use np.core.umath.clip if available
In NumPy versions >=1.17, `np.clip` got significantly slower.  The best
workaround currently seems to be to use the internal
`np.core.umath.clip` instead (see
https://github.com/numpy/numpy/issues/14281).
`np.core.umath.clip` does not exist in versions <1.17 so in this case we
fall back to `np.clip` (which is okay because in these versions
`np.clip` was still good).