Set initial position based on finger type
For each finger type specify its default initial position in the
finger_types_data module (using the same values as in the configuration
of the real robots).
This is ensures that the fingers are not stuck in the ground upon
initialisation, which is the case for some of the robots if initialised
to all-zero positions.