fix: Use np.min to avoid mypy error.
Mypy complained about this because the input is a numpy type rather than
a native Python type.
It was still working as expected but since changing `min` to `np.min` is
trivial and makes the type checker happy, let's simply do that.
This should not change the behaviour of the code in any way.